+++ src/Sylius/Bundle/CoreBundle/spec/Sylius/Bundle/CoreBundle/Twig/RestrictedZoneExtensionSpec.php
@@ -1,0 +1,33 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Paweł Jędrzejewski
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace spec\Sylius\Bundle\CoreBundle\Twig;
+
+use PhpSpec\ObjectBehavior;
+use Sylius\Bundle\CoreBundle\Templating\Helper\RestrictedZoneHelper;
+
+class RestrictedZoneExtensionSpec extends ObjectBehavior
+{
+    function let(RestrictedZoneHelper $helper)
+    {
+        $this->beConstructedWith($helper);
+    }
+
+    function it_is_initializable()
+    {
+        $this->shouldHaveType('Sylius\Bundle\CoreBundle\Twig\RestrictedZoneExtension');
+    }
+
+    function it_is_a_twig_extension()
+    {
+        $this->shouldHaveType('Twig_Extension');
+    }
+}