+++ src/Sylius/Bundle/WebBundle/DependencyInjection/SyliusWebExtension.php
@@ -30,5 +30,6 @@ class SyliusWebExtension extends Extension
     {
         $loader = new XmlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('services.xml');
+        $loader->load('twig.xml');
     }
 }