+++ src/Sylius/Bundle/InventoryBundle/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+### v0.10.0
+
+* Twig extension was renamed from `SyliusInventoryExtension` into `InventoryExtension`,
+  also the service name was changed from `sylius.inventory_twig` to `sylius.twig.extension.inventory`.
+
 ### v0.9.0
 
 * Release before components.